refactor modifiders
