+++ src/request/mod.rs
@@ -1,13 +1,8 @@
-mod auth_header;
-mod body;
 mod cors;
-mod custom_headers;
 mod dispatcher;
 mod error;
-mod headers;
 mod modifier;
 mod response;
-mod timeout;
 
 pub use cors::same_origin_redirect_policy;
 pub use dispatcher::Dispatcher;
