+++ src/request/modifier/body.rs
@@ -1,7 +1,7 @@
 use reqwest::RequestBuilder;
 
-use super::modifier::RequestModifier;
 use super::RequestError;
+use super::RequestModifier;
 use crate::oauth2::OAuth2Config;
 use crate::options::Opts;
 pub struct Body {}
