+++ src/request/modifier/custom_headers.rs
@@ -1,9 +1,9 @@
 use reqwest::header::{CONTENT_TYPE, USER_AGENT};
 use reqwest::RequestBuilder;
 
-use super::error::RequestError;
 use super::headers::Headers;
-use super::modifier::RequestModifier;
+use super::RequestError;
+use super::RequestModifier;
 use crate::oauth2::OAuth2Config;
 use crate::options::Opts;
 use crate::version;
