+++ src/request/modifier/mod.rs
@@ -1,11 +1,17 @@
+mod auth_header;
+mod body;
+mod custom_headers;
+mod headers;
+mod timeout;
+
 use crate::oauth2::{AccessToken, OAuth2Config};
 use crate::options::Opts;
 use crate::request::error::RequestError;
 
-use super::auth_header::CustomAuthHeader;
-use super::body::Body;
-use super::custom_headers::CustomHeaders;
-use super::timeout::Timeout;
+use self::auth_header::CustomAuthHeader;
+use self::body::Body;
+use self::custom_headers::CustomHeaders;
+use self::timeout::Timeout;
 use reqwest::RequestBuilder;
 
 pub trait RequestModifier {
